Moved `ShareComponent` to `/Posts` folder
