+++ src/components/Posts/ShareComponent.vue
